+++ sandpack-react/src/components/CodeEditor/CodeMirror.tsx
@@ -45,6 +45,7 @@ import {
 import { useSyntaxHighlight } from "./useSyntaxHighlight";
 import {
   getCodeMirrorLanguage,
+  getLanguageFromFile,
   getEditorTheme,
   getSyntaxHighlight,
   useCombinedRefs,
@@ -85,6 +86,9 @@ interface CodeMirrorProps {
    * appears when `readOnly` is `true`
    */
   showReadOnly?: boolean;
+  /**
+   * Provides a way to draw or style a piece of the content.
+   */
   decorators?: Decorators;
   initMode: SandpackInitMode;
   id?: string;
@@ -152,7 +156,8 @@ export const CodeMirror = React.forwardRef<CodeMirrorRef, CodeMirrorProps>(
       }
     }, [initMode, isIntersecting]);
 
-    const langSupport = getCodeMirrorLanguage(filePath, fileType);
+    const languageExtension = getLanguageFromFile(filePath, fileType);
+    const langSupport = getCodeMirrorLanguage(languageExtension);
     const highlightTheme = getSyntaxHighlight(theme);
 
     const syntaxHighlightRender = useSyntaxHighlight({
@@ -376,7 +381,7 @@ export const CodeMirror = React.forwardRef<CodeMirrorRef, CodeMirrorProps>(
           <pre
             ref={combinedRef}
             className={classNames(
-              c("cm", editorState),
+              c("cm", editorState, languageExtension),
               cmClassName,
               tokensClassName
             )}
@@ -408,7 +413,7 @@ export const CodeMirror = React.forwardRef<CodeMirrorRef, CodeMirrorProps>(
           filePath ? `Code Editor for ${getFileName(filePath)}` : `Code Editor`
         }
         className={classNames(
-          c("cm", editorState),
+          c("cm", editorState, languageExtension),
           cmClassName,
           editorClassName,
           tokensClassName