+++ sandpack-react/src/components/Preview/index.tsx
@@ -137,7 +137,6 @@ export const SandpackPreview = ({
       customStyle={{
         ...customStyle,
         ...viewportStyle,
-        display: status !== "idle" ? "flex" : "none",
       }}
     >
       {showNavigator ? (